+++ extras/doc/CHANGELOG.md
@@ -2,3 +2,6 @@
 
 * Version 1.0.0 August 2025
 	* first release
+* Version 1.0.1 May 2026
+	* Fixed compiler warning for 'rxdata'
+	* Added variable I2C and SPI speed.